fix: reduce Reconcile cyclomatic complexity to pass gocyclo lint
Extract gateway status application and status building into helper
methods in mcpserver_controller_gateway.go, bringing Reconcile
complexity from 35 down to 30 (the gocyclo threshold).
